resolve circular dependancy with compare_state
